requirements.txt
python-dotenv
playwright
requests
openai
fastapi
uvicorn[standard]
jinja2
aiofiles
python-socks
apscheduler
httpx[socks]
Pillow
pyzbar
qrcode
pytest
pytest-asyncio
coverage
opencv-python
pytesseract
numpy
